Refactor: Rename "index.ts" to "middleware.ts"
